+++ examples/vue-next-webpack-preview-master/src/App.vue
@@ -2,6 +2,8 @@
   <img src="./logo.png" />
   <h1>Hello Vue 3!</h1>
 
+  <p>hydrating: {{ hydrating }}</p>
+
   <ul>
     <li
       v-for="({ name, use }, i) in components"
@@ -36,6 +38,8 @@ import Swapi from "./components/SWAPI.vue";
 import TodoList from "./components/TodoList.vue";
 import ScrollLock from "./components/ScrollLock.vue";
 import { ref } from "vue";
+import { useHydration } from "vue-composable";
+import { onMounted } from "vue";
 
 export default {
   components: {
@@ -51,6 +55,13 @@ export default {
   },
 
   setup(props, ctx) {
+    const hydrating = useHydration();
+
+    console.log("setup hydrating", hydrating.value);
+    onMounted(() => {
+      console.log("mounted hydrating", hydrating.value);
+    });
+
     const selected = ref(0);
     const components = require
       .context("./components", true, /\.(vue)$/)
@@ -65,7 +76,8 @@ export default {
 
     return {
       components,
-      selected
+      selected,
+      hydrating
     };
   }
 };