+++ examples/vue-next-webpack-preview-master/src/main.js
@@ -1,4 +1,9 @@
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from "vue";
+import App from "./App.vue";
+import { hydrationPlugin } from "vue-composable";
 
-createApp(App).mount('#app')
+const app = createApp(App);
+
+app.use(hydrationPlugin);
+
+app.mount("#app");