+++ packages/global.d.ts
@@ -4,3 +4,4 @@ declare var __JSDOM__: boolean;
 declare var __COMMIT__: string;
 declare var __VERSION__: string;
 declare var __VUE_2__: boolean;
+declare var __SSR__: boolean;