+++ packages/vue-composable/src/index.ts
@@ -12,7 +12,11 @@ export * from "./date";
 export * from "./format";
 export * from "./validation";
 export * from "./i18n";
+export * from "./meta";
+export * from "./ssr";
+export * from "./state";
 
 export const VERSION = __VERSION__;
+// istanbul ignore next
 export const VUE_VERSION: "2" | "3" = __VUE_2__ ? "2" : "3";
 export const COMMIT = __COMMIT__;