+++ packages/vue-composable/src/misc/index.ts
@@ -1,3 +1,5 @@
 export * from "./matchMedia";
 export * from "./sharedRef";
 export * from "./lockScroll";
+export * from "./vmodel";
+export * from "./injectFactory";